LiveGuard/LiveGuard_backend/requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
setuptools==75.1.0
sqlparse==0.5.1
